[WIP] Discovery of new deps for llvm-rocm
Identified new dependencies for LLVM. This will require a bit of
refactoring. LLVM appears to now depend on the HSA runtime, which
(if I understand correctly) is bundled as part of rocr-runtime.
LLVM also explicity refers to roct-thunk-interface in its dependencies,
so I need to figure out that one too.